+++ CHANGELOG.md
@@ -21,6 +21,7 @@ A more detailed list of changes is available in the corresponding milestones for
   - **[com.google.fonts/check/008]:** com.google.fonts/check/underline_thickness
   - **[com.google.fonts/check/009]:** com.google.fonts/check/panose_proportion
   - **[com.google.fonts/check/010]:** com.google.fonts/check/panose_familytype
+  - **[com.google.fonts/check/011]:** com.google.fonts/check/equal_numbers_of_glyphs
   - **[com.google.fonts/check/035]:** com.google.fonts/check/ftxvalidator
   - **[com.google.fonts/check/036]:** com.google.fonts/check/ots
   - **[com.google.fonts/check/038]:** com.google.fonts/check/fontforge_stderr